refactor state into smaller files
